requirements.txt
Cython
EasyDict
opencv-python
mxnet-cu80
